UserApp/src/main/java/com/javahunter/UserApp/controller/UserController.java
package com.javahunter.UserApp.controller;

import com.javahunter.UserApp.entity.User;
import com.javahunter.UserApp.service.UserService;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@Slf4j
@RestController
@RequiredArgsConstructor
@RequestMapping("/user")
public class UserController {

    private final UserService userService;
    @GetMapping("/hello")
    public String hello(){
        return "Hey...You made it...!!!... Congratulations";
    }

    @PostMapping("/save")
    public ResponseEntity<User> saveUser(@RequestBody User user) {
        return new ResponseEntity<>(userService.saveUser(user), HttpStatus.OK);
    }
}
